20181104_2

+++ src/services/permissions.service.js
@@ -1,9 +1,11 @@
 const PermissionItem = require("../dto/permission.item");
 const debuger = require('debug');
+const mongoose = require('mongoose');
 const permissionServiceDebuger = debuger('app:permissionService');
 const permissionModel = PermissionItem.model;
 var _ = require('lodash');
 const NoSuchMediaException = require("../errors/no-such-media.error");
+const IllegalArgumentError = require('../errors/illegal-argument.error');
 
 class PermissionService {
     async addPermission(permission) {
@@ -12,13 +14,15 @@ class PermissionService {
         try {
             return await model.save();
         } catch(e) {
-            permissionServiceDebuger(e.message);
+            permissionServiceDebuger(e);
             throw e;
         }
     }    
     async updatePermission(permission) {
         const permissionJson = permission.toJson();
         try{
+            if(!mongoose.Types.ObjectId.isValid(permissionJson.id)) 
+                throw new IllegalArgumentError('ID');
             const result = await permissionModel.findByIdAndUpdate(permissionJson.id, 
                 _.pick(permissionJson, ["resource", "method", "description"]),
                 { new: true });
@@ -31,8 +35,10 @@ class PermissionService {
     }
     async removePermission(id) {
         try {
+            if(!mongoose.Types.ObjectId.isValid(id)) 
+                throw new IllegalArgumentError('ID');
             const result = await permissionModel.findByIdAndRemove(id);
-            if(!result) throw new NoSuchMediaException("Permission with the given ID was not found.")
+            if(!result) throw new NoSuchMediaException("Permission")
             return result;
         } catch(e) {
             permissionServiceDebuger(e);
@@ -41,9 +47,11 @@ class PermissionService {
     }
     async getPermission(id) {
         try {
+            if(!mongoose.Types.ObjectId.isValid(id)) 
+                throw new IllegalArgumentError('ID');
             const result = await permissionModel.findById(id);
             if(!result) {
-                throw new NoSuchMediaException("Permission with the given ID was not found.", "400")
+                throw new NoSuchMediaException("Permission");
             }
             return result;
         } catch(e) {
@@ -55,7 +63,7 @@ class PermissionService {
             const result = await permissionModel.find();
             return result;
         } catch(e) {
-            permissionServiceDebuger(e.message);
+            permissionServiceDebuger(e);
             throw e;
         }
     }
